+++ src/components/sections/CircuitSelect/index.tsx
@@ -34,88 +34,10 @@ import {
     LitinskiCompilationOptions,
 } from "../../../lib/compilationOptions"
 import { Slices } from "../../../lib/slices"
+import CompilationOptionsSelector from "./CompilationOptionsSelector"
 
 
-interface CompilationOptionsSelectorProps {
-    compilationOptions: CompilationOptions
-    setCompilationOptions: React.Dispatch<CompilationOptions>
-}
 
-const CompilationOptionsSelector = ({compilationOptions, setCompilationOptions} : CompilationOptionsSelectorProps) => {
-
-    const updateFastSlicerOptions = (fastSlicerOptions : FastSlicerOptions) =>
-        setCompilationOptions({...compilationOptions, fastSlicerOptions})
-    const updateLitinskiCompilationOptions= (litinskiCompilationOptions : LitinskiCompilationOptions) =>
-        setCompilationOptions({...compilationOptions, litinskiCompilationOptions})
-
-    return <VStack>
-        <Select>
-            <option
-                onClick={() => setCompilationOptions({ ...compilationOptions, kind: "FastSlicerOptions" })}>
-                Fast Slicer with direct compilation via ZX
-            </option>
-            <option
-                onClick={() => setCompilationOptions({ ...compilationOptions, kind: "LitinskiCompilationOptions" })}>
-                Litinski style compilation via Pauli Rotations
-            </option>
-        </Select>
-        {(compilationOptions.kind === "FastSlicerOptions")
-            ?
-            <>
-                <Text>
-                    <b>Note:</b> If submitting .qasm, the fast slicer only supports
-                    {' '}
-                    <Link
-                        color="teal.500"
-                        href="https://github.com/latticesurgery-com/liblsqecc#qasm-support-highly-experimental/"
-                        target="_blank"
-                        isExternal
-                    >
-                        a very small subset
-                    </Link>.
-                </Text>
-                <HStack>
-                    <Text>Layout:</Text>
-                    <Select defaultValue={LayoutType[compilationOptions.fastSlicerOptions.layoutType]}>
-                        {Object.keys(LayoutType).filter((v) => isNaN(Number(v))).map((item) =>
-                            <option
-                                key={item}
-                                onClick={() => updateFastSlicerOptions({
-                                    ...compilationOptions.fastSlicerOptions, layoutType : LayoutType[item as keyof typeof LayoutType]
-                                })}>
-                                {item}
-                            </option>
-                        )
-                        }
-                    </Select>
-                    <Text>Apply corrective terms:</Text>
-                    <Select defaultValue={CorrectiveTermBehaviour[compilationOptions.fastSlicerOptions.correctiveTermBehaviour]}>
-                        {Object.keys(CorrectiveTermBehaviour).filter((v) => isNaN(Number(v))).map((item) =>
-                            <option
-                                key={item}
-                                onClick={() => updateFastSlicerOptions({
-                                    ...compilationOptions.fastSlicerOptions, correctiveTermBehaviour : CorrectiveTermBehaviour[item as keyof typeof CorrectiveTermBehaviour]
-                                })}>
-                                {item}
-                            </option>
-                        )
-                        }
-                    </Select>
-                </HStack>
-            </>
-            :
-            <Checkbox
-                isChecked={compilationOptions.litinskiCompilationOptions.doStabilizerCommutingTransform}
-                onChange={(e) => updateLitinskiCompilationOptions({
-                            ...compilationOptions.litinskiCompilationOptions,
-                            doStabilizerCommutingTransform: e.target.checked
-                        })}
-            >
-                <Text as={"span"}>Pi/4 removing transform</Text>
-            </Checkbox>
-        }
-    </VStack>
-}
 
 interface EmscriptenSlicerResult {
     err : string
@@ -128,8 +50,7 @@ const runCompilation = async (
     setAppState:React.Dispatch<AppState>,
     fileContents : string,
     compilationOptions:CompilationOptions,
-    extension: string,
-    repeats: number) => {
+    extension: string) => {
 
     // Modify State on Compile Request Submit
     setAppState({
@@ -138,7 +59,7 @@ const runCompilation = async (
     })
 
     if (compilationOptions.kind == "LitinskiCompilationOptions")
-        submitCompileRequest(setAppState, fileContents, compilationOptions.litinskiCompilationOptions, repeats)
+        submitCompileRequest(setAppState, fileContents, compilationOptions.litinskiCompilationOptions)
     else {
         const {fastSlicerOptions} = compilationOptions;
 
@@ -204,7 +125,7 @@ type CircuitSelectProps = {
     setRepeats: (value: number) => void
 }
 
-const CircuitSelect = ({ appState, setAppState, repeats, setRepeats }: CircuitSelectProps) => {
+const CircuitSelect = ({ appState, setAppState }: CircuitSelectProps) => {
     const [compilationOptions, setCompilationOptions] = useState<CompilationOptions>(defaultCompilationOptions)
     const readFile = (file: File) => {
         return new Promise((resolve, reject) => {
@@ -235,7 +156,7 @@ const CircuitSelect = ({ appState, setAppState, repeats, setRepeats }: CircuitSe
                 })
             }
         } else {
-            await runCompilation(setAppState, data as string, compilationOptions, extension, repeats)
+            await runCompilation(setAppState, data as string, compilationOptions, extension)
         }
     }
 
@@ -243,7 +164,7 @@ const CircuitSelect = ({ appState, setAppState, repeats, setRepeats }: CircuitSe
         const file_url = `${process.env.PUBLIC_URL}/assets/demo_circuits/${example}`
         const data = await fetch(file_url).then((response) => response.text())
         if (data) {
-            await runCompilation(setAppState, data as string, compilationOptions, "qasm", repeats)
+            await runCompilation(setAppState, data as string, compilationOptions, "qasm")
         }
     }
 
@@ -275,27 +196,6 @@ const CircuitSelect = ({ appState, setAppState, repeats, setRepeats }: CircuitSe
                 isLoading={appState.compilationIsLoading}
             />
             <CompilationOptionsSelector compilationOptions={compilationOptions} setCompilationOptions={setCompilationOptions} />
-            {isDevMode() && (
-                <Box p="3" rounded="lg" borderWidth="3px" borderColor="#98ff98">
-                    <Flex gap="2">
-                        <Text margin="auto">Repeats</Text>
-                        <NumberInput
-                            w="120px"
-                            defaultValue={repeats}
-                            onChange={(value) => {
-                                setRepeats(parseInt(value))
-                            }}
-                            isDisabled={appState.compilationIsLoading}
-                        >
-                            <NumberInputField />
-                            <NumberInputStepper>
-                                <NumberIncrementStepper />
-                                <NumberDecrementStepper />
-                            </NumberInputStepper>
-                        </NumberInput>
-                    </Flex>
-                </Box>
-            )}
         </Flex>
     )
 }
